pyproject.toml
[tool.poetry]
name = "battleship-revamp-retro"
version = "1.0.0"
description = "Retro Battleship game with FastAPI, HTMX, and modern DevOps."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
fastapi = ">=0.111,<1.0"
uvicorn = { extras = ["standard"], version = ">=0.30,<1.0" }
httpx = ">=0.27,<0.28"
jinja2 = ">=3.1,<4.0"
python-multipart = ">=0.0.12,<0.1"
fastapi-sso = ">=0.15.0"
psycopg = { version = ">=3.2,<4.0", extras = ["binary"] }
sqlalchemy = ">=2.0,<3.0"
passlib = { extras = ["argon2"], version = ">=1.7,<2.0" }
email-validator = ">=2.1,<3.0"
python-jose = { extras = ["cryptography"], version = ">=3.3,<4.0" }
python-decouple = ">=3.8,<4.0"
pydantic = ">=2.6,<3.0"
ecdsa = ">=0.19.2"

[tool.poetry.group.dev.dependencies]
ruff = "0.5.6"
pytest = "8.4.1"
pytest-asyncio = "1.1.0"
pytest-cov = "5.0.0"
coverage = { extras = ["toml"], version = "7.6.1" }
pre-commit = "4.0.1"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "S"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["S101"]
"src/battleship/auth/schemas.py" = ["S105"]

[tool.ruff.lint.isort]
known-first-party = ["src"]
